Postgres sql

+++ src/main/java/com/educator/repository/CourseRepository.java
@@ -4,9 +4,7 @@ import com.educator.domain.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Long> {
-    Course findAllById(Long id);
+public interface CourseRepository extends JpaRepository<Course, Integer> {
+    Course findAllById(int id);
 }
